# api.yml
openapi: '3.0.4'
info:
  title: Sports Facilities Management System
  description: |-
    Esta es la definición de la API que se va a usar dentro del sistema de gestión de centros deportivos para poder hacer uso de todas las funciones que ofrece. Estas pueden ser separadas en función del tipo de usuario final que va a poder acceder a ellas:
    - Administradores: gestión de instalaciones, dispositivos y usuarios.
    - Clientes: gestión del perfil y reserva de instalaciones.
  contact:
    email: pabloquesadarojo@gmail.com
  version: '1.3.0'
servers:
  - url: https://api.server.test/
security:
  - BearerAuth: []
...
paths:
...
  /complexes:
    get:
      tags:
        - complexes
      summary: Obtiene una lista de complejos deportivos del sistema.
      description: Obtiene una lista de complejos deportivos del sistema que puede estar filtrada en función de una serie de parámetros.
      operationId: getComplexes
      x-security-roles:
        - ADMIN
        - USER
      responses:
        '200':
          description: Se devuelve correctamente la lista de complejos deportivos.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComplexInfo'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - complexes
      summary: Crea un nuevo complejo deportivo en el sistema.
      description: Crea un nuevo complejo deportivo en el sistema, asignándole un identificador único y completando la información asociada a este.
      operationId: createComplex
      x-security-roles:
        - SUPERADMIN
      requestBody:
        description: Objeto que contiene la información del complejo deportivo a crear.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplexInfo'
        required: true
      responses:
        '200':
          description: El complejo deportivo se ha creado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplexInfo'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /complexes/{complexId}/courts/{courtId}/status:
    parameters:
      - $ref: '#/components/parameters/complexId'
      - $ref: '#/components/parameters/courtId'
    get:
      tags:
        - courts
      summary: Obtiene el estado de una pista.
      description: Obtiene el estado de una pista concreta.
      operationId: getCourtStatus
      x-security-roles:
        - ADMIN
        - USER
      responses:
        '200':
          description: Se devuelve correctamente el estado de la pista pedida.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourtStatus'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No se ha encontrado la pista.
          content:
            application/json: {}
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - courts
      summary: Actualiza el estado de una pista.
      description: Actualiza el estado de una pista concreta.
      operationId: updateCourtStatus
      x-security-roles:
        - ADMIN
      requestBody:
        description: Objeto que contiene la información actualizada del estado de la pista.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourtStatus'
        required: true
      responses:
        '200':
          description: El estado de la pista se ha actualizado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourtStatus'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No se ha encontrado la pista.
          content:
            application/json: {}
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /complexes/{complexId}/courts/{courtId}/devices:
    parameters:
      - $ref: '#/components/parameters/complexId'
      - $ref: '#/components/parameters/courtId'
    get:
      tags:
        - courts
      summary: Obtiene una lista de pistas del sistema.
      description: Obtiene una lista de pistas del sistema que puede estar filtrada en función de una serie de parámetros.
      operationId: getCourtDevices
      x-security-roles:
        - ADMIN
      responses:
        '200':
          description: Se devuelve correctamente la lista de pistas.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourtInfo'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /complexes/{complexId}/devices/{deviceId}/courts:
    parameters:
      - $ref: '#/components/parameters/complexId'
      - $ref: '#/components/parameters/deviceId'
    get:
      tags:
        - devices
      summary: Obtiene las pistas asociadas a un dispositivo.
      description: Obtiene una lista de las pistas asociadas a un dispositivo concreto.
      operationId: getDeviceCourts
      x-security-roles:
        - ADMIN
      responses:
        '200':
          description: Se devuelve correctamente la lista de pistas asociadas al dispositivo pedido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceCourts'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No se ha encontrado el dispositivo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - devices
      summary: Asocia pistas a un dispositivo.
      description: Asocia una o varias pistas a un dispositivo concreto.
      operationId: addDeviceCourts
      x-security-roles:
        - ADMIN
      requestBody:
        description: Objeto que contiene la lista de identificadores de las pistas a asociar al dispositivo.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCourts'
        required: true
      responses:
        '201':
          description: Las pistas se han asociado correctamente al dispositivo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceCourts'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No se ha encontrado el dispositivo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          description: Código de estado HTTP.
        message:
          type: string
          description: Mensaje descriptivo del error.
        error:
          type: string
          description: Tipo de error (ej. Bad Request, Unauthorized).
...
    WeatherData:
      type: object
      properties:
        temperature:
          type: number
          description: Temperatura actual.
        precip_intensity_prev:
          type: number
          description: Precipitación registrada en la hora previa.
        precip_intensity_curr:
          type: number
          description: Precipitación actual.
        precip_probability_curr:
          type: integer
          description: Probabilidad de precipitación actual.
        precip_probability_next:
          type: integer
          description: Probabilidad de precipitación para la próxima hora.
      description: Datos meteorológicos obtenidos de la API de OpenMeteo.
...
    DeviceCourts:
      type: array
      items:
        type: integer
...
    TimeInfo:
      type: object
      properties:
        timeIni:
          type: string
          format: time
        timeEnd:
          type: string
          format: time
    ComplexInfo:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        time:
          $ref: '#/components/schemas/TimeInfo'
        locLongitude:
          type: number
          format: double
        locLatitude:
          type: number
          format: double
        weather:
          $ref: '#/components/schemas/WeatherData'
    CourtSport:
      type: string
      enum:
        - football
        - basketball
        - tennis
        - padel
    CourtInfo:
      type: object
      properties:
        id:
          type: integer
        sport:
          $ref: '#/components/schemas/CourtSport'
        name:
          type: string
        description:
          type: string
        maxPeople:
          type: integer
    CourtStatus:
      type: string
      enum:
        - open
        - maintenance
        - blocked
        - weather
      default: open
...
  parameters:
    complexId:
      name: complexId
      in: path
      required: true
      description: Identificador del sistema.
      schema:
        type: integer
    deviceId:
      name: deviceId
      in: path
      required: true
      description: Identificador del dispositivo.
      schema:
        type: integer
...
    courtId:
      name: courtId
      in: path
      required: true
      description: Identificador de la pista.
      schema:
        type: integer
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    DeviceAuth:
      type: apiKey
      in: header
      name: X-Device-Token
      description: Token de autenticación del dispositivo, utilizado para autenticar las peticiones de los dispositivos IoT.